requirements/dev.txt
-r base.txt

django-extensions==3.2.1
